refactor: share connector account target keys
